api_keyProxyを使用してAPIキーの読み込みを更新し、CharcterAISettingProxyを追加して設定の読み込み・更新処理を移行
